Servidor/requirements.txt
flask-restful
pymongo
pyjwt
bcrypt
pyopenssl
python-dotenv

# For deployment
gunicorn